update ui tests
