+++ src/utils/select.test.tsx
@@ -1,0 +1,8 @@
+import highlightText from './select';
+
+it('can highlight text', () => {
+    const t = 'this is a test';
+    const q = 's is';
+
+    highlightText(t, q);
+});
