+++ src/utils/select.tsx
@@ -1,0 +1,38 @@
+/* eslint-disable no-constant-condition */
+/* eslint-disable no-useless-escape */
+import React from 'react';
+
+function escapeRegExpChars(text: string) {
+    return text.replace(/([.*+?^=!:${}()|\[\]\/\\])/g, '\\$1');
+}
+
+export default function highlightText(text: string, query: string) {
+    let lastIndex = 0;
+    const words = query
+        .split(/\s+/)
+        .filter((word) => word.length > 0)
+        .map(escapeRegExpChars);
+    if (words.length === 0) {
+        return [text];
+    }
+    const regexp = new RegExp(words.join('|'), 'gi');
+    const tokens: React.ReactNode[] = [];
+    while (true) {
+        const match = regexp.exec(text);
+        if (!match) {
+            break;
+        }
+        const { length } = match[0];
+        const before = text.slice(lastIndex, regexp.lastIndex - length);
+        if (before.length > 0) {
+            tokens.push(before);
+        }
+        lastIndex = regexp.lastIndex;
+        tokens.push(<strong key={lastIndex}>{match[0]}</strong>);
+    }
+    const rest = text.slice(lastIndex);
+    if (rest.length > 0) {
+        tokens.push(rest);
+    }
+    return tokens;
+}
